Created context processor for base.html navbar
