+++ landing/views.py
@@ -1,48 +1,25 @@
 import random
 
-from enum import Enum
-
 from django.db.models import Avg
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render
 
-from account.models.client import Client
-from account.models.profile import Profile
+from landing.context_processors import user_context
 from landing.models import TeamMember
 from account.models.professional import Professional
 
 MAX_PROFESSIONALS = 48
 
 
-class UserType(Enum):
-    CLIENT = 'C'
-    PROFESSIONAL = 'P'
-
-
 def homepage(request):
-    user = request.user
-    user_id = None
-
-    if user.is_authenticated:
-        profile = get_object_or_404(Profile, user_id=user)
-
-        if profile.user_type == UserType['PROFESSIONAL'].value:
-            user_id = get_object_or_404(Professional, profile_id=profile)
-        elif profile.user_type == UserType.CLIENT.value:
-            user_id = get_object_or_404(Client, profile_id=profile)
-
     # TODO: add images for professionals
     db_professionals = Professional.objects.annotate(avg_rating=Avg('review__rating'))
-    # Rounds average rating for each professional
     round_avg_rating(db_professionals)
-    # Shuffle professionals randomly
     professionals = list(db_professionals)
     random.shuffle(professionals)
-    # Context
     context = {
-        'user_id': user_id,
-        'user_type': UserType.__members__,
-        'professionals': professionals[:MAX_PROFESSIONALS],  # Shows only the 'MAX_PROFESSIONALS' first professionals
+        'professionals': professionals[:MAX_PROFESSIONALS],
     }
+    context.update(user_context(request))
     return render(request, 'landing/homepage.html', context=context)
 
 
@@ -58,5 +35,8 @@ def round_avg_rating(professionals):
 
 
 def learn_more(request):
-    context = {'team_members': TeamMember.objects.all()}
+    context = {
+        'team_members': TeamMember.objects.all(),
+    }
+    context.update(user_context(request))
     return render(request, 'landing/learn-more.html', context=context)
